\name{fillgrowthdata}
\alias{fillgrowthdata}
\title{Minimal doc}
\usage{
  fillgrowthdata(date, data, growthdata)
}
\description{
  this function fills in the growth data with the data that
  is being read in. It is important to realize that we need
  to have a data structure that keeps track of time. This
  is a really sparse structure, but it allows us keep time
  proportional.
}

