\name{rectplot}
\alias{rectplot}
\title{Minimal doc}
\usage{
  rectplot(ser, bins, xmin, xmax, ylim, barcol1, barcol2)
}
\description{
  This file makes all the length frequency and peak trough
  plots. It consists of two functions. rectplot and
  rqFreqPlot. It was written by Aaron Greenberg for the
  Project New ELEFAN These are the fancy plots that are
  super important to ELEFAN
}

